Now generated code could be in separate module (use ModuleName config param to set module name) Disabled blueprint support for enums to support uint16/32 enums Added native delegates that could be used for c++ code Added "OnContextStateChanged" delegate to handle Added "OnShutdown" delegate Added "OnServiceGrpcError" delegate for handling grpc errors Added bIsEnded for client streaming response to handle stream ends Context handle now shared ReleaseService now can be forced without timeout

+++ Tools/generate_code.cmd
@@ -66,29 +66,41 @@ if !errorFlag! equ 1 (
     exit /b 1
 )
 
+set ADDITIONAL_INCLUDE=
+if "%3" NEQ "" set ADDITIONAL_INCLUDE= --proto_path="%3"
 ::call protoc.exe
 "%PROTOC_EXE_PATH%" ^
- --proto_path="%PROTOBUF_INC_PATH%" --proto_path="%INPUT_PROTO_PATH%" ^
+ --proto_path="%PROTOBUF_INC_PATH%" --proto_path="%INPUT_PROTO_PATH%" %ADDITIONAL_INCLUDE%^
  --cpp_out="%CPP_OUTPUT_PATH%" ^
  --plugin=protoc-gen-grpc="%GRPC_CPP_PLUGIN_EXE_PATH%" --grpc_out=%CPP_OUTPUT_PATH% ^
  --plugin=protoc-gen-turbolink="%TURBOLINK_PLUGIN_PATH%" --turbolink_out="%OUTPUT_PATH%" ^
  --turbolink_opt="GenerateJsonCode=true" ^
- %INPUT_PROTO_FILE%
+ %INPUT_PROTO_PATH%\%INPUT_PROTO_FILE%
 
-:: fix protobuf compile warning 
+if %ERRORLEVEL% NEQ 0 exit /b %ERRORLEVEL
+
+:: fix protobuf compile warning
 call :FixCompileWarning "%FIX_PROTO_H%" %CPP_OUTPUT_PATH%\%INPUT_PROTO_FILE% "pb.h"
 call :FixCompileWarning "%FIX_PROTO_CPP%" %CPP_OUTPUT_PATH%\%INPUT_PROTO_FILE% "pb.cc"
 goto :eof
 
 :FixCompileWarning
 set FIX_FILE=%1
-set FILE_PATH=%~p2
+set FILE_PATH=%~dp2
 set FILE_NAME=%~n2.%~3
 
 pushd %FILE_PATH%
-copy /b %FIX_FILE%+%FILE_NAME% %FILE_NAME%.tmp
-del /f %FILE_NAME%
-rename %FILE_NAME%.tmp %FILE_NAME%
-popd 
+if "%~n2%" EQU "*" (
+    for %%f in (%FILE_NAME%) do (
+        copy /b %FIX_FILE%+%%f %%f.tmp
+        del /f %%f
+        rename %%f.tmp %%f
+    )
+) else (
+    copy /b %FIX_FILE%+%FILE_NAME% %FILE_NAME%.tmp
+    del /f %FILE_NAME%
+    rename %FILE_NAME%.tmp %FILE_NAME%
+)
+popd
 
 goto :eof
